create middleware to check user role organise folder structure
