+++ src/middlewares/index.js
@@ -1,0 +1,4 @@
+module.exports = {
+    authMiddleware: require("./middleware"),
+    checkPermissions: require("./checkPermissions")
+}
