+++ static/roles.js
@@ -1,0 +1,12 @@
+module.exports = {
+    admin : {
+        update_parcel_status: true,
+        get_all_parcels: true,
+    },
+    buyer : {
+        get_parcels: true,
+        update_destination: true,
+        get_single_parcel: true,
+        create_delivery_order: true
+    }
+}
